fix definition of atiyah-hirzebruch spectral sequence, define serre spectral sequence
The construction of the Serre spectral sequence is done up to 11 sorry's, all which are marked with 'TODO FOR SSS'. 8 of them are equivalences related to cohomology (6 of which are corollaries of the other 2), 2 of them are calculations on int, and the last is in the definition of a spectrum map.