proof naturality of pointed funext. This finishes the proof of the Serre Spectral Sequence.
We use a different proof strategy for the naturality than pursued the last week.
We proof the unpointed version of the naturality by generalizing it from loops to paths so that we can apply path induction.
For the pointed version, we do some ugly calculations to cancel noncomputable applications of funext